Amend WaitForTimelineChange action to allow waiting for arbitrary timeline.
In some situations, the timeline can't be specified because it is created
internally by the media source under test. If the test still needs to wait for
a timeline update, this change allows to do that by specifying an expected
timeline of null.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=190768386