Make ABR playback tests more resilient to HostActivity failures.
We currently fail the entire playback tests if we exceed a start timeout
or the test gets stopped forcefully. This is problematic as the start timeout
may happen from time to time for the ABR tests and we'd still like to have the
remaining playbacks.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=222406639